+++ extensions/amazon-aws/src/components/common/action.tsx
@@ -1,4 +1,5 @@
-import { Action, ActionPanel, Keyboard } from "@raycast/api";
+import { Action, ActionPanel, Icon, Keyboard } from "@raycast/api";
+import { getEnumKeysExcludingCurrent } from "../../util";
 
 export class AwsAction {
   public static Console = ({ url }: { url: string }) => (
@@ -20,4 +21,35 @@ export class AwsAction {
       shortcut={{ modifiers: ["opt"], key: "e" }}
     />
   );
+
+  public static SwitchResourceType<T extends object>({
+    setResourceType,
+    current,
+    enumType,
+  }: {
+    setResourceType: (resourceType: T[keyof T]) => void;
+    current: T[keyof T];
+    enumType: T;
+  }) {
+    const allKeysExceptCurrent = getEnumKeysExcludingCurrent(enumType, current);
+    return (
+      <ActionPanel.Section>
+        <ActionPanel.Submenu
+          title="Switch Resource Type"
+          shortcut={{ modifiers: ["cmd", "shift"], key: "r" }}
+          filtering
+          icon={Icon.Box}
+        >
+          {allKeysExceptCurrent.map((k) => (
+            <Action
+              title={`Show ${k.toString()}`}
+              onAction={() => setResourceType(enumType[k])}
+              key={k.toString()}
+              icon={Icon.Box}
+            />
+          ))}
+        </ActionPanel.Submenu>
+      </ActionPanel.Section>
+    );
+  }
 }