+++ extensions/amazon-aws/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - CodePipeline: Adds action to retry failed stage, stop ongoing execution and toggle stage transition
 - Secrets Manager: Retrieves secret value in a new Detail navigation and adds extra metadata
+- Cloudformation: Improves the resource switch action
 
 ## [SQS, CFN, DDB, CW Logs commands] - 2024-06-14